requirements.txt
flask>=2.3.0
python-telegram-bot>=20.0,<21.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
httpx>=0.24.0,<0.26.0
